Allow classify's main function to be imported
This is required to allow it be run from inside a PEX bundle.
